chore: split into packages
