+++ justfile
@@ -1,6 +1,6 @@
 # Run CLI (builds first)
 run *commands:
-    pnpm run build --silent && node ./dist/index.mjs {{commands}}
+    pnpm -r run build --silent && node packages/cli/dist/index.mjs {{commands}}
 
 # Alias for run
 alias r := run
@@ -11,8 +11,8 @@ local-deploy: build-icrc21-canister
 
 # Build ICRC-21 test canister WASM
 build-icrc21-canister:
-    mise exec -- cargo build --manifest-path tests/icrc21-canister/Cargo.toml --target wasm32-unknown-unknown --release
-    cp tests/icrc21-canister/target/wasm32-unknown-unknown/release/icrc21_canister.wasm tests/icrc21-canister/icrc21_canister.wasm
+    mise exec -- cargo build --manifest-path packages/icrc21-agent/tests/icrc21-canister/Cargo.toml --target wasm32-unknown-unknown --release
+    cp packages/icrc21-agent/tests/icrc21-canister/target/wasm32-unknown-unknown/release/icrc21_canister.wasm packages/icrc21-agent/tests/icrc21-canister/icrc21_canister.wasm
 
 test:
     pnpm test
