+++ package.json
@@ -1,25 +1,12 @@
 {
-  "name": "@dfinity/hardware-wallet-cli",
-  "version": "0.6.0",
-  "description": "A CLI to interact with the Internet Computer App on Ledger Nano S/X devices.",
+  "private": true,
   "type": "module",
-  "main": "./dist/index.mjs",
-  "files": [
-    "dist/index.mjs",
-    "README.md",
-    "LICENSE"
-  ],
   "scripts": {
     "format": "prettier --write .",
     "pretest": "just build-icrc21-canister",
     "test": "vitest run",
-    "typecheck": "tsc --noEmit --skipLibCheck",
-    "build": "node rmdir.mjs && node esbuild.mjs",
-    "prepack": "pnpm run build",
-    "clean": "pnpm run build",
-    "refresh": "rm -rf ./node_modules ./package-lock.json && pnpm install",
-    "execute": "pnpm run build && node ./dist/index.mjs",
-    "run": "node ./dist/index.mjs",
+    "typecheck": "pnpm -r run typecheck",
+    "build": "pnpm -r run build",
     "update:icp-sdk": "pnpm update @icp-sdk/canisters @icp-sdk/core @dfinity/utils",
     "build:tokens": "ts-node ./scripts/get-tokens-list.ts"
   },
@@ -31,36 +18,15 @@
     "type": "git",
     "url": "git+https://github.com/dfinity/hardware-wallet-cli.git"
   },
-  "author": "",
   "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/dfinity/hardware-wallet-cli/issues"
-  },
-  "homepage": "https://github.com/dfinity/hardware-wallet-cli#readme",
-  "dependencies": {
-    "@dfinity/utils": "^4.1.0",
-    "@icp-sdk/canisters": "^3.2.0",
-    "@icp-sdk/core": "^5.0.0",
-    "@ledgerhq/hw-transport-node-hid-noevents": "^6.3.0",
-    "@ledgerhq/hw-transport-webhid": "^6.27.1",
-    "@zondax/ledger-icp": "^3.2.6",
-    "chalk": "^4.1.2",
-    "commander": "^9.0.0",
-    "node-hid": "^3.2.0",
-    "node-window-polyfill": "^1.0.2"
-  },
   "devDependencies": {
     "@dfinity/pic": "^0.21.0",
+    "@icp-sdk/core": "^5.0.0",
     "@types/google-protobuf": "^3.15.6",
-    "@types/node": "^17.0.16",
-    "@types/node-hid": "^1.3.1",
-    "esbuild": "0.25.0",
     "prettier": "^2.6.2",
     "ts-node": "^10.9.1",
+    "typescript": "^5.0.0",
     "vitest": "^4.1.2"
-  },
-  "bin": {
-    "ic-hardware-wallet": "./dist/index.mjs"
   },
   "overrides": {
     "semver": "^7.5.3"
