+++ packages/cli/rmdir.mjs
