+++ packages/icrc21-agent/src/index.ts
@@ -1,0 +1,2 @@
+export { Icrc21Agent } from "./icrc21-agent";
+export type { Icrc21Identity } from "./icrc21-identity";
