+++ packages/icrc21-agent/tests/icrc21-canister/idl.ts
