+++ packages/icrc21-agent/tsconfig.tsbuildinfo
@@ -1,0 +1,1 @@
+{"root":["./src/icrc21-agent.ts","./src/icrc21-identity.ts","./src/index.ts"],"version":"5.9.3"}
