+++ packages/ledger-hw-identity/src/index.ts
@@ -1,0 +1,2 @@
+export { LedgerIdentity } from "./identity";
+export { Secp256k1PublicKey } from "./secp256k1";
