+++ packages/ledger-hw-identity/tsconfig.tsbuildinfo
@@ -1,0 +1,1 @@
+{"root":["./src/identity.ts","./src/index.ts","./src/secp256k1.ts"],"version":"5.9.3"}
